# artisan/doorman/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Transfer (pull, verify, scan, resign and push) artefacts between repositories",
        "title": "Artisan's Doorman",
        "contact": {
            "name": "gatblau",
            "url": "http://onix.gatblau.org/",
            "email": "onix@gatblau.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.4"
    },
    "paths": {
...
        "/command": {
            "put": {
                "description": "creates or updates a command",
                "consumes": [
                    "application/yaml",
                    " application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Commands"
                ],
                "summary": "Creates or updates a command",
                "parameters": [
                    {
                        "description": "the data for the command to persist",
                        "name": "key",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Command"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/event/{service-id}/{bucket-name}/{folder-name}": {
            "post": {
                "description": "Triggers the ingestion of a specification",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Triggers the ingestion of an artisan spec artefacts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "a unique identifier for the bucket endpoint (e.g. x-minio-deployment-id for MinIO)",
                        "name": "deployment-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the name of the bucket that contains the uploaded files",
                        "name": "bucket-name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the name of the folder within the bucket that contains the uploaded files",
                        "name": "folder-name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/key": {
            "put": {
                "description": "creates or updates a cryptographic key used by either inbound or outbound routes to verify or sign\npackages respectively",
                "consumes": [
                    "application/yaml",
                    " application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Keys"
                ],
                "summary": "Creates or updates a cryptographic key",
                "parameters": [
                    {
                        "description": "the data for the key to persist",
                        "name": "key",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notification": {
            "get": {
                "description": "gets all notifications",
                "produces": [
                    "application/json",
                    " application/yaml",
                    " application/xml"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Gets all notifications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "creates or updates a notification",
                "consumes": [
                    "application/yaml",
                    " application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Creates or updates a notification",
                "parameters": [
                    {
                        "description": "the data for the notification to persist",
                        "name": "key",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Notification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notification-template": {
            "get": {
                "description": "gets all notification templates",
                "produces": [
                    "application/json",
                    " application/yaml",
                    " application/xml"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Gets all notification templates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "creates or updates a notification template",
                "consumes": [
                    "application/yaml",
                    " application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Creates or updates a notification template",
                "parameters": [
                    {
                        "description": "the data for the notification template to persist",
                        "name": "key",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.NotificationTemplate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/pipe/{name}": {
            "get": {
                "description": "gets a pipeline",
                "produces": [
                    "application/json",
                    " application/yaml",
                    " application/xml"
                ],
                "tags": [
                    "Pipelines"
                ],
                "summary": "Gets a pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the name of the pipeline to retrieve",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.Command": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "the command description",
                    "type": "string",
                    "example": "scan files in specified path"
                },
                "error_regex": {
                    "description": "a regex used to determine if the command execution has errored",
                    "type": "string",
                    "example": ".*Infected files: [^0].*"
                },
                "name": {
                    "description": "a unique name for the command",
                    "type": "string",
                    "example": "clamscan"
                },
                "stopOnError": {
                    "description": "determines if the process should stop on a command execution error",
                    "type": "boolean",
                    "example": true
                },
                "value": {
                    "description": "the value of the command",
                    "type": "string",
                    "example": "freshclam \u0026\u0026 clamscan -r ${path}"
                }
            }
        },
...
        "types.NotificationTemplate": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content of the template",
                    "type": "string"
                },
                "name": {
                    "description": "Name unique identifier for notification template",
                    "type": "string"
                },
                "subject": {
                    "description": "Subject of the notification",
                    "type": "string"
                }
            }
        },
...
    }
}